Split up tubesync/sync/views.py
