+++ tubesync/sync/urls.py
@@ -2,9 +2,10 @@ from django.urls import path
 from .views import (DashboardView, SourcesView, ValidateSourceView, AddSourceView,
                     SourceView, UpdateSourceView, DeleteSourceView, MediaView,
                     MediaThumbView, MediaItemView, MediaRedownloadView, MediaSkipView,
-                    MediaEnableView, MediaContent, TasksView, CompletedTasksView, ResetTasks,
-                    TaskScheduleView, MediaServersView, AddMediaServerView, MediaServerView,
-                    DeleteMediaServerView, UpdateMediaServerView)
+                    MediaEnableView, MediaContent, TasksView, CompletedTasksView,
+                    ResetTasks, TaskScheduleView, MediaServersView, AddMediaServerView,
+                    MediaServerView, DeleteMediaServerView, UpdateMediaServerView,
+                    RevokeTaskView, SourceSyncNowView, )
 
 
 app_name = 'sync'
@@ -36,7 +37,7 @@ urlpatterns = [
 
     path(
         'source-sync-now/<uuid:pk>',
-        SourcesView.as_view(),
+        SourceSyncNowView.as_view(),
         name='source-sync-now',
     ),
 
@@ -124,7 +125,7 @@ urlpatterns = [
 
     path(
         'task/<int:pk>/cancel',
-        TasksView.as_view(),
+        RevokeTaskView.as_view(),
         name='revoke-task',
     ),
