+++ tubesync/sync/views/__init__.py
@@ -1,0 +1,25 @@
+from .dashboard import DashboardView
+from .sources import (SourcesView, SourceSyncNowView, ValidateSourceView,
+                      AddSourceView, SourceView, UpdateSourceView,
+                      DeleteSourceView)
+from .media import (MediaView, MediaThumbView, MediaItemView,
+                    MediaRedownloadView, MediaSkipView, MediaEnableView,
+                    MediaContent)
+from .tasks import (TasksView, RevokeTaskView, CompletedTasksView,
+                    ResetTasks, TaskScheduleView)
+from .mediaservers import (MediaServersView, AddMediaServerView,
+                           MediaServerView, DeleteMediaServerView,
+                           UpdateMediaServerView)
+
+__all__ = [
+    'DashboardView',
+    'SourcesView', 'SourceSyncNowView', 'ValidateSourceView', 'AddSourceView',
+    'SourceView', 'UpdateSourceView', 'DeleteSourceView',
+    'MediaView', 'MediaThumbView', 'MediaItemView',
+    'MediaRedownloadView', 'MediaSkipView', 'MediaEnableView',
+    'MediaContent',
+    'TasksView', 'RevokeTaskView', 'CompletedTasksView', 'ResetTasks',
+    'TaskScheduleView',
+    'MediaServersView', 'AddMediaServerView', 'MediaServerView',
+    'DeleteMediaServerView', 'UpdateMediaServerView',
+]
